+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## StreamChat
 ### 🐞 Fixed
 - Fix pin/unpin updates not propagated in `LivestreamChannelController` when the message is not in the local messages list [#4032](https://github.com/GetStream/stream-chat-swift/pull/4032)
+- Fix `ChatChannel.latestMessages` including thread replies [#4024](https://github.com/GetStream/stream-chat-swift/pull/4024)
+### 🔄 Changed
+- Remove `ChatChannel.previewMessage` in favour of computing preview from `latestMessages` [#4024](https://github.com/GetStream/stream-chat-swift/pull/4024)
 
 ## StreamChatUI
+### ✅ Added
+- Show deleted messages in channel list preview [#4024](https://github.com/GetStream/stream-chat-swift/pull/4024)
 ### 🐞 Fixed
 - Fix clearing unread messages on channel open when the first unread message is not visible [#4013](https://github.com/GetStream/stream-chat-swift/pull/4013)